+++ core/components/svgsanitizer/lexicon/en/setting.inc.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * Lexicon entries for svgSanitizer system settings
+ *
+ * @package svgsanitizer
+ * @subpackage lexicon
+ */
+
+$_lang['area_caching'] = 'Caching';
+$_lang['setting_svgsanitizer.cache_key'] = 'Cache Key';
+$_lang['setting_svgsanitizer.cache_key_desc'] = 'A key to identify the cache handler. If you want to clear the SVG cache together with the MODX cache (not recommended), set this to \'default\'.';
+$_lang['setting_svgsanitizer.cache_expires'] = 'Cache Expiry';
+$_lang['setting_svgsanitizer.cache_expires_desc'] = 'The amount of time in seconds, that the sanitized SVG will be cached.';